Handle mega menu footer link accessibility when menu is constructed.
